+++ CHANGELOG.md
@@ -7,9 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 - Added `from_bytes_truncating_at_nul` to `CString`
-- Added `into_bytes` to `CString`
-- Added `into_bytes_with_nul` to `CString`
-- Added `into_string` to `CString`
+- Added `CString::{into_bytes, into_bytes_with_nul, into_string}`
 
 ## [v0.9.2] 2025-11-12