+++ CHANGELOG.md
@@ -13,6 +13,7 @@ of panicking drop implementations.
 - Added `CString::{into_bytes, into_bytes_with_nul, into_string}`
 - Added `pop_front_if` and `pop_back_if` to `Deque`
 - Made `Vec::from_array` const.
+- Fixed long division being instroduced by the const-erasure in spsc
 
 ## [v0.9.2] 2025-11-12